+++ envvar.go
@@ -23,10 +23,11 @@ const (
 
 // EnvVarSet is a set of defined environment variables.
 type EnvVarSet struct {
-	parsed        bool
-	vars          map[string]*EnvVar
-	errorHandling ErrorHandling
-	output        io.Writer // nil means stderr; use out() accessor
+	parsed            bool
+	vars              map[NormalizedName]*EnvVar
+	errorHandling     ErrorHandling
+	output            io.Writer // nil means stderr; use out() accessor
+	normalizeNameFunc NormalizeFunc
 }
 
 // EnvVar represents the state of an environment variable.
@@ -67,10 +68,10 @@ func (s *EnvVarSet) Var(value Value, name string, usage string) {
 // AddEnvVar will add the environment variable to the EnvVarSet.
 func (s *EnvVarSet) AddEnvVar(envVar *EnvVar) {
 	if s.vars == nil {
-		s.vars = make(map[string]*EnvVar)
+		s.vars = make(map[NormalizedName]*EnvVar)
 	}
 
-	name := envVar.Name
+	name := s.normalizeVarName(envVar.Name)
 
 	_, alreadyThere := s.vars[name]
 	if alreadyThere {
@@ -104,8 +105,8 @@ func (s *EnvVarSet) SetOutput(output io.Writer) {
 func (s *EnvVarSet) Parse(environment map[string]string) error {
 	s.parsed = true
 
-	for key, value := range environment {
-		if ev, ok := s.vars[key]; ok {
+	for name, value := range environment {
+		if ev, ok := s.vars[NormalizedName(name)]; ok {
 			err := ev.Value.Set(value)
 			if err != nil {
 				switch s.errorHandling {